Move fetchLicenseEvidence to BomBuilder
